+++ source/timemory/components/timing.hpp
@@ -60,17 +60,17 @@ struct real_clock : public base<real_clock>
     static std::string display_unit() { return "sec"; }
     static value_type  record() { return tim::get_clock_real_now<int64_t, ratio_t>(); }
 
-    double compute_display() const
+    float compute_display() const
     {
         auto val = (is_transient) ? accum : value;
-        return static_cast<double>(val / static_cast<double>(ratio_t::den) *
-                                   base_type::get_unit());
+        return static_cast<float>(val / static_cast<float>(ratio_t::den) *
+                                  base_type::get_unit());
     }
 
-    double get() const
+    float get() const
     {
         auto val = (is_transient) ? accum : value;
-        return static_cast<double>(val / static_cast<double>(ratio_t::den));
+        return static_cast<float>(val / static_cast<float>(ratio_t::den));
     }
 
     void start()
@@ -116,11 +116,11 @@ struct system_clock : public base<system_clock>
     static std::string descript() { return "system time"; }
     static std::string display_unit() { return "sec"; }
     static value_type  record() { return tim::get_clock_system_now<int64_t, ratio_t>(); }
-    double             compute_display() const
+    float              compute_display() const
     {
         auto val = (is_transient) ? accum : value;
-        return static_cast<double>(val / static_cast<double>(ratio_t::den) *
-                                   base_type::get_unit());
+        return static_cast<float>(val / static_cast<float>(ratio_t::den) *
+                                  base_type::get_unit());
     }
     void start()
     {
@@ -158,11 +158,11 @@ struct user_clock : public base<user_clock>
     static std::string descript() { return "user time"; }
     static std::string display_unit() { return "sec"; }
     static value_type  record() { return tim::get_clock_user_now<int64_t, ratio_t>(); }
-    double             compute_display() const
+    float              compute_display() const
     {
         auto val = (is_transient) ? accum : value;
-        return static_cast<double>(val / static_cast<double>(ratio_t::den) *
-                                   base_type::get_unit());
+        return static_cast<float>(val / static_cast<float>(ratio_t::den) *
+                                  base_type::get_unit());
     }
     void start()
     {
@@ -200,11 +200,11 @@ struct cpu_clock : public base<cpu_clock>
     static std::string descript() { return "cpu time"; }
     static std::string display_unit() { return "sec"; }
     static value_type  record() { return tim::get_clock_cpu_now<int64_t, ratio_t>(); }
-    double             compute_display() const
+    float              compute_display() const
     {
         auto val = (is_transient) ? accum : value;
-        return static_cast<double>(val / static_cast<double>(ratio_t::den) *
-                                   base_type::get_unit());
+        return static_cast<float>(val / static_cast<float>(ratio_t::den) *
+                                  base_type::get_unit());
     }
     void start()
     {
@@ -242,11 +242,11 @@ struct monotonic_clock : public base<monotonic_clock>
     {
         return tim::get_clock_monotonic_now<int64_t, ratio_t>();
     }
-    double compute_display() const
+    float compute_display() const
     {
         auto val = (is_transient) ? accum : value;
-        return static_cast<double>(val / static_cast<double>(ratio_t::den) *
-                                   base_type::get_unit());
+        return static_cast<float>(val / static_cast<float>(ratio_t::den) *
+                                  base_type::get_unit());
     }
     void start()
     {
@@ -285,11 +285,11 @@ struct monotonic_raw_clock : public base<monotonic_raw_clock>
     {
         return tim::get_clock_monotonic_raw_now<int64_t, ratio_t>();
     }
-    double compute_display() const
+    float compute_display() const
     {
         auto val = (is_transient) ? accum : value;
-        return static_cast<double>(val / static_cast<double>(ratio_t::den) *
-                                   base_type::get_unit());
+        return static_cast<float>(val / static_cast<float>(ratio_t::den) *
+                                  base_type::get_unit());
     }
     void start()
     {
@@ -326,11 +326,11 @@ struct thread_cpu_clock : public base<thread_cpu_clock>
     static std::string descript() { return "thread cpu time"; }
     static std::string display_unit() { return "sec"; }
     static value_type  record() { return tim::get_clock_thread_now<int64_t, ratio_t>(); }
-    double             compute_display() const
+    float              compute_display() const
     {
         auto val = (is_transient) ? accum : value;
-        return static_cast<double>(val / static_cast<double>(ratio_t::den) *
-                                   base_type::get_unit());
+        return static_cast<float>(val / static_cast<float>(ratio_t::den) *
+                                  base_type::get_unit());
     }
     void start()
     {
@@ -366,11 +366,11 @@ struct process_cpu_clock : public base<process_cpu_clock>
     static std::string descript() { return "process cpu time"; }
     static std::string display_unit() { return "sec"; }
     static value_type  record() { return tim::get_clock_process_now<int64_t, ratio_t>(); }
-    double             compute_display() const
+    float              compute_display() const
     {
         auto val = (is_transient) ? accum : value;
-        return static_cast<double>(val / static_cast<double>(ratio_t::den) *
-                                   base_type::get_unit());
+        return static_cast<float>(val / static_cast<float>(ratio_t::den) *
+                                  base_type::get_unit());
     }
     void start()
     {
@@ -414,16 +414,16 @@ struct cpu_util : public base<cpu_util, std::pair<int64_t, int64_t>>
     {
         return value_type(cpu_clock::record(), real_clock::record());
     }
-    double compute_display() const
+    float compute_display() const
     {
-        double denom =
+        float denom =
             (accum.second > 0) ? accum.second : ((value.second > 0) ? value.second : 1);
-        double numer =
+        float numer =
             (accum.second > 0) ? accum.first : ((value.second > 0) ? value.first : 0);
-        return 100.0 * static_cast<double>(numer) / static_cast<double>(denom);
+        return 100.0 * static_cast<float>(numer) / static_cast<float>(denom);
     }
-    double serialization() { return compute_display(); }
-    void   start()
+    float serialization() { return compute_display(); }
+    void  start()
     {
         set_started();
         value = record();
@@ -486,16 +486,16 @@ struct process_cpu_util : public base<process_cpu_util, std::pair<int64_t, int64
     {
         return value_type(process_cpu_clock::record(), real_clock::record());
     }
-    double compute_display() const
+    float compute_display() const
     {
-        double denom =
+        float denom =
             (accum.second > 0) ? accum.second : ((value.second > 0) ? value.second : 1);
-        double numer =
+        float numer =
             (accum.second > 0) ? accum.first : ((value.second > 0) ? value.first : 0);
-        return 100.0 * static_cast<double>(numer) / static_cast<double>(denom);
+        return 100.0 * static_cast<float>(numer) / static_cast<float>(denom);
     }
-    double serialization() { return compute_display(); }
-    void   start()
+    float serialization() { return compute_display(); }
+    void  start()
     {
         set_started();
         value = record();
@@ -558,16 +558,16 @@ struct thread_cpu_util : public base<thread_cpu_util, std::pair<int64_t, int64_t
     {
         return value_type(thread_cpu_clock::record(), real_clock::record());
     }
-    double compute_display() const
+    float compute_display() const
     {
-        double denom =
+        float denom =
             (accum.second > 0) ? accum.second : ((value.second > 0) ? value.second : 1);
-        double numer =
+        float numer =
             (accum.second > 0) ? accum.first : ((value.second > 0) ? value.first : 0);
-        return 100.0 * static_cast<double>(numer) / static_cast<double>(denom);
+        return 100.0 * static_cast<float>(numer) / static_cast<float>(denom);
     }
-    double serialization() { return compute_display(); }
-    void   start()
+    float serialization() { return compute_display(); }
+    void  start()
     {
         set_started();
         value = record();
